separated comments list with input
