+++ src/app/domains/Comment/components/forms/CommentSimpleForm/CommentSimpleForm.style.js
@@ -1,0 +1,17 @@
+import { makeStyles } from '@material-ui/core/styles'
+
+export const useStyles = makeStyles({
+  root: {
+    padding: '2px 4px',
+    display: 'flex',
+    alignItems: 'center'
+  },
+  input: {
+    marginLeft: '10px',
+    flex: 1
+  },
+  divider: {
+    height: 28,
+    margin: 4
+  }
+})
