+++ src/app/domains/Comment/components/forms/CommentSimpleForm/__tests__/CommentSimpleForm.test.js
@@ -1,0 +1,7 @@
+import React from 'react'
+import { render } from '@testing-library/react'
+import CommentSimpleForm from '../CommentSimpleForm.template'
+
+test('Render CommentSimpleForm test.', () => {
+  render(<CommentSimpleForm />)
+})
