+++ src/app/domains/Comment/components/forms/CommentSimpleForm/index.js
@@ -1,0 +1,1 @@
+export { default as CommentSimpleForm } from './CommentSimpleForm.template'
