--- src/app/domains/Comment/components/forms/CommentSimpleView/CommentSimpleView.template.js
@@ -1,8 +1,0 @@
-const CommentSimpleView = (props) => {
-  return <div>CommentSimpleView</div>
-}
-
-CommentSimpleView.propTypes = {}
-CommentSimpleView.defaultProps = {}
-
-export default CommentSimpleView
