--- src/app/domains/Comment/components/forms/CommentSimpleView/__tests__/CommentSimpleView.test.js
@@ -1,7 +1,0 @@
-import React from 'react'
-import { render } from '@testing-library/react'
-import CommentSimpleView from '../CommentSimpleView.template'
-
-test('Render CommentSimpleView test.', () => {
-  render(<CommentSimpleView />)
-})
