--- src/app/domains/Comment/components/forms/CommentSimpleView/index.js
@@ -1,1 +1,0 @@
-export { default as CommentSimpleView } from './CommentSimpleView.template'
